refactor(vendored): migrate ex-WezTerm crates to upgraded dependency APIs
Part of the 2026-06-14/15 workspace dependency-upgrade batch. Adapts the
vendored FrankenTerm (ex-WezTerm) crates under `frankenterm/` to the new
upstream-dependency APIs whose versions were bumped in the preceding
manifest commit. Manual, per-call-site ports only (no script-based
rewrites, per AGENTS.md).

Migration highlights by dependency:

  - mlua: drop the explicit `'lua` lifetime parameter from
    `UserData`/`UserDataFields`/`UserDataMethods` impls and FromLua/IntoLua
    call sites (mux-lua domain/pane/tab/window/lib, termwiz-funcs,
    url-funcs, config/lua.rs, config/config.rs).
  - bitflags: adapt to the new generated API (`.bits()`, `from_bits_*`,
    iteration) in config/font, config/keyassignment, input-types,
    ssh/sftp/types, window/lib.
  - csscolorparser / colorgrad: color-parse and gradient-builder API
    updates in color-types, config/background, term/terminalstate/sixel.
  - ordered-float: `NotNan`/`OrderedFloat` `.into_inner()` and
    `new_f32`-style constructors in surface and term image scaling.
  - syn: derive-macro parse-surface updates in config/derive/attr and
    dynamic/derive/attr.
  - governor: rate-limiter quota/clock surface in ratelim.
  - criterion / misc bench deps: `black_box` and harness-API updates
    across the char-props, config, rangeset, scripting, termwiz, and
    window benches; lfucache and window/bitmaps internal adjustments.

Also folds in the rustfmt/clippy churn the new toolchain/dep versions
surfaced in these files (import ordering, expression reflow). No behavior
change is intended; these are compile-compatibility ports.

+++ frankenterm/term/src/terminalstate/sixel.rs
@@ -102,13 +102,13 @@ impl TerminalState {
                     // go from sixel red to standard hsl red.
                     // Negative values wrap around the circle.
                     // https://github.com/wezterm/wezterm/issues/775
-                    let angle = (*hue_angle as f64) - 120.0;
-                    let angle = if angle < 0. { 360.0 + angle } else { angle };
+                    let angle = (*hue_angle as f32) - 120.0;
+                    let angle = if angle < 0.0 { 360.0 + angle } else { angle };
                     let c = csscolorparser::Color::from_hsla(
                         angle,
-                        *saturation as f64 / 100.,
-                        *lightness as f64 / 100.,
-                        1.,
+                        *saturation as f32 / 100.0,
+                        *lightness as f32 / 100.0,
+                        1.0,
                     );
                     let [r, g, b, _] = c.to_rgba8();
                     if color_map.len() >= self.config.max_color_map_entries()
